Create any_base_to_any_base.java
I had done a mistake in this question. the mistake was that the function for converting the value in decimal or any base declare after the class ended. so the error is shown like enum or class not defined error. when I did correct it then our program is successfully run.