Create diff_of_ARRAY.java
in this question, I usually forget to apply the upgrade condition of IDX in the loop so the same number, again and again, is printing. so always remember one thing if the same number is, again and again, is printing and run time error or exceed may be you forget to upgradation so that that error Is coming.